requirements.txt
discord.py
datetime
mcrcon
urllib3
requests
mctools